+++ superduper/backends/base/cdc.py
@@ -27,4 +27,3 @@ class CDCBackend(BaseBackend):
         :param value: ``Datalayer`` instance.
         """
         self._db = value
-        self.initialize()