+++ CHANGELOG.md
@@ -97,6 +97,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix Qdrant collection name
 - Fix the ordering and sequencing of jobs initiated on `db.apply`
 - Fix rest routes with db injection and prefix.
+- Fix cluster db setter.
 
 ## [0.3.0](https://github.com/superduper-io/superduper/compare/0.3.0...0.2.0])    (2024-Jun-21)